+++ tests/testthat/test-10-themingTests.R
@@ -193,6 +193,7 @@ for(i in 1:nrow(scenarios)) {
 
 scenarios <- testScenarios("theming tests:  styling basics 1")
 for(i in 1:nrow(scenarios)) {
+  if(!isDevelopmentVersion) break
   evaluationMode <- scenarios$evaluationMode[i]
   processingLibrary <- scenarios$processingLibrary[i]
   description <- scenarios$description[i]
@@ -223,6 +224,7 @@ for(i in 1:nrow(scenarios)) {
 
 scenarios <- testScenarios("theming tests:  styling basics 2")
 for(i in 1:nrow(scenarios)) {
+  if(!isDevelopmentVersion) break
   evaluationMode <- scenarios$evaluationMode[i]
   processingLibrary <- scenarios$processingLibrary[i]
   description <- scenarios$description[i]