+++ tests/testthat/test-25-pivotRowColumns.R
@@ -373,4 +373,94 @@ for(i in 1:nrow(scenarios)) {
 }
 
 
+scenarios <- testScenarios("pvt row/col tests:  vectorised methods")
+for(i in 1:nrow(scenarios)) {
+  evaluationMode <- scenarios$evaluationMode[i]
+  processingLibrary <- scenarios$processingLibrary[i]
+  description <- scenarios$description[i]
+  countFunction <- scenarios$countFunction[i]
+
+  test_that(description, {
+
+    library(pivottabler)
+    pt <- PivotTable$new(processingLibrary=processingLibrary, evaluationMode=evaluationMode, compatibility=list(noDataGroupNBSP=TRUE))
+    pt$addData(bhmtrains)
+    pt$addColumnDataGroups("TrainCategory")
+    pt$addColumnDataGroups("PowerType")
+    pt$addRowDataGroups("TOC")
+    pt$addRowDataGroups("Status")
+    pt$defineCalculation(calculationName="TotalTrains", summariseExpression=countFunction)
+    pt$evaluatePivot()
+
+    # pt$renderPivot()
+    # sum(pt$cells$asMatrix(), na.rm=TRUE)
+    # prepStr(as.character(pt$getHtml()))
+    html <- "<table class=\"Table\">\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"2\" colspan=\"2\">&nbsp;</th>\n    <th class=\"ColumnHeader\" colspan=\"4\">Express Passenger</th>\n    <th class=\"ColumnHeader\" colspan=\"3\">Ordinary Passenger</th>\n    <th class=\"ColumnHeader\">Total</th>\n  </tr>\n  <tr>\n    <th class=\"ColumnHeader\">DMU</th>\n    <th class=\"ColumnHeader\">EMU</th>\n    <th class=\"ColumnHeader\">HST</th>\n    <th class=\"ColumnHeader\">Total</th>\n    <th class=\"ColumnHeader\">DMU</th>\n    <th class=\"ColumnHeader\">EMU</th>\n    <th class=\"ColumnHeader\">Total</th>\n    <th class=\"ColumnHeader\"></th>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"4\">Arriva Trains Wales</th>\n    <th class=\"RowHeader\">A</th>\n    <td class=\"Cell\">3018</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">3018</td>\n    <td class=\"Cell\">815</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">815</td>\n    <td class=\"Total\">3833</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">C</th>\n    <td class=\"Cell\">59</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">59</td>\n    <td class=\"Cell\">15</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">15</td>\n    <td class=\"Total\">74</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">R</th>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">2</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">2</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">Total</th>\n    <td class=\"Total\">3079</td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">3079</td>\n    <td class=\"Total\">830</td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">830</td>\n    <td class=\"Total\">3909</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"4\">CrossCountry</th>\n    <th class=\"RowHeader\">A</th>\n    <td class=\"Cell\">21561</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\">709</td>\n    <td class=\"Total\">22270</td>\n    <td class=\"Cell\">60</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">60</td>\n    <td class=\"Total\">22330</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">C</th>\n    <td class=\"Cell\">546</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\">23</td>\n    <td class=\"Total\">569</td>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">2</td>\n    <td class=\"Total\">571</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">R</th>\n    <td class=\"Cell\">26</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">26</td>\n    <td class=\"Cell\">1</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">1</td>\n    <td class=\"Total\">27</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">Total</th>\n    <td class=\"Total\">22133</td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">732</td>\n    <td class=\"Total\">22865</td>\n    <td class=\"Total\">63</td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">63</td>\n    <td class=\"Total\">22928</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"4\">London Midland</th>\n    <th class=\"RowHeader\">A</th>\n    <td class=\"Cell\">5534</td>\n    <td class=\"Cell\">8599</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">14133</td>\n    <td class=\"Cell\">5520</td>\n    <td class=\"Cell\">27331</td>\n    <td class=\"Total\">32851</td>\n    <td class=\"Total\">46984</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">C</th>\n    <td class=\"Cell\">101</td>\n    <td class=\"Cell\">235</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">336</td>\n    <td class=\"Cell\">67</td>\n    <td class=\"Cell\">847</td>\n    <td class=\"Total\">914</td>\n    <td class=\"Total\">1250</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">R</th>\n    <td class=\"Cell\">3</td>\n    <td class=\"Cell\">15</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">18</td>\n    <td class=\"Cell\">4</td>\n    <td class=\"Cell\">23</td>\n    <td class=\"Total\">27</td>\n    <td class=\"Total\">45</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">Total</th>\n    <td class=\"Total\">5638</td>\n    <td class=\"Total\">8849</td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">14487</td>\n    <td class=\"Total\">5591</td>\n    <td class=\"Total\">28201</td>\n    <td class=\"Total\">33792</td>\n    <td class=\"Total\">48279</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\" rowspan=\"4\">Virgin Trains</th>\n    <th class=\"RowHeader\">A</th>\n    <td class=\"Cell\">2028</td>\n    <td class=\"Cell\">6331</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">8359</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">8359</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">C</th>\n    <td class=\"Cell\">107</td>\n    <td class=\"Cell\">119</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">226</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">226</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">R</th>\n    <td class=\"Cell\">2</td>\n    <td class=\"Cell\">7</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\">9</td>\n    <td class=\"Cell\"></td>\n    <td class=\"Cell\"></td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">9</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">Total</th>\n    <td class=\"Total\">2137</td>\n    <td class=\"Total\">6457</td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">8594</td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\"></td>\n    <td class=\"Total\">8594</td>\n  </tr>\n  <tr>\n    <th class=\"RowHeader\">Total</th>\n    <th class=\"RowHeader\"></th>\n    <td class=\"Total\">32987</td>\n    <td class=\"Total\">15306</td>\n    <td class=\"Total\">732</td>\n    <td class=\"Total\">49025</td>\n    <td class=\"Total\">6484</td>\n    <td class=\"Total\">28201</td>\n    <td class=\"Total\">34685</td>\n    <td class=\"Total\">83710</td>\n  </tr>\n</table>"
+
+    expect_equal(sum(pt$cells$asMatrix(), na.rm=TRUE), 753390)
+    expect_identical(as.character(pt$getHtml()), html)
+
+    fx <- function(x) { x$caption }
+    gx <- function(x) {
+      captions <- lapply(unlist(x), fx)
+      paste(captions, collapse=" ")
+    }
+
+    grps <- pt$getColumnGroupsByLevel(c(1, 2))
+    text <- "Express Passenger Ordinary Passenger Total DMU EMU HST Total DMU EMU Total "
+    expect_equal(length(grps), 2)
+    expect_identical(gx(grps), text)
+
+    grps <- pt$getColumnGroupsByLevel(c(1, 2), collapse=TRUE)
+    text <- "Express Passenger Ordinary Passenger Total DMU EMU HST Total DMU EMU Total "
+    expect_equal(length(grps), 11)
+    expect_identical(gx(grps), text)
+
+    levels <- pt$findGroupColumnNumbers(grps, collapse=TRUE)
+    expect_equal(length(levels), 8)
+    expect_equal(sum(levels), 36)
+
+    grps <- pt$getLeafColumnGroup(c(2, 4, 7))
+    text <- "EMU Total Total"
+    expect_equal(length(grps), 3)
+    expect_identical(gx(grps), text)
+
+    levels <- pt$findGroupColumnNumbers(grps)
+    expect_equal(length(levels), 3)
+    expect_equal(sum(unlist(levels)), 13)
+
+    levels <- pt$findGroupColumnNumbers(grps, collapse=TRUE)
+    expect_equal(length(levels), 3)
+    expect_equal(sum(levels), 13)
+
+    grps <- pt$getRowGroupsByLevel(c(1, 2))
+    text <- "Arriva Trains Wales CrossCountry London Midland Virgin Trains Total A C R Total A C R Total A C R Total A C R Total "
+    expect_equal(length(grps), 2)
+    expect_identical(gx(grps), text)
+
+    grps <- pt$getRowGroupsByLevel(c(1, 2), collapse=TRUE)
+    text <- "Arriva Trains Wales CrossCountry London Midland Virgin Trains Total A C R Total A C R Total A C R Total A C R Total "
+    expect_equal(length(grps), 22)
+    expect_identical(gx(grps), text)
+
+    levels <- pt$findGroupRowNumbers(grps, collapse=TRUE)
+    expect_equal(length(levels), 17)
+    expect_equal(sum(levels), 153)
+
+    grps <- pt$getLeafRowGroup(c(2, 4, 7))
+    text <- "C Total R"
+    expect_equal(length(grps), 3)
+    expect_identical(gx(grps), text)
+
+    levels <- pt$findGroupRowNumbers(grps)
+    expect_equal(length(levels), 3)
+    expect_equal(sum(unlist(levels)), 13)
+
+    levels <- pt$findGroupRowNumbers(grps, collapse=TRUE)
+    expect_equal(length(levels), 3)
+    expect_equal(sum(levels), 13)
+  })
+}
+
+